This will implement redux-loader interface #80
When Redux-Loaders are present it will register the 5 loaders with it.
Redux is able to handle train-cargo un-/loading. It will not change any
if there is not the Redux mod installed. Remember to enable train cargo
loading in mod-settings (ingame menu) if you want the loaders to load
trains!